sessionコントローラー、ヘルパー、ビューを戻す
