+++ app/controllers/users_controller.rb
@@ -1,8 +1,6 @@
 # frozen_string_literal: true
 
 class UsersController < ApplicationController
-  before_action :redirect_logged_in_user_to_root_page, only: %i[new create]
-
   def index
     @users = User.all
   end
